requirements.txt
google-generativeai
openai
python-dotenv
bs4
playwright

